+++ app/src/main/java/mx/itesm/pachontli/navigationDrawer/NavigationDrawerActivity.kt
@@ -38,7 +38,7 @@ class NavigationDrawerActivity : AppCompatActivity() {
         // menu should be considered as top level destinations.
         appBarConfiguration = AppBarConfiguration(
             setOf(
-                R.id.nav_home, R.id.nav_gallery, R.id.nav_slideshow
+                R.id.nav_home, R.id.nav_veterinarios, R.id.nav_interactua
             ), drawerLayout
         )
         setupActionBarWithNavController(navController, appBarConfiguration)
