+++ app/src/main/java/mx/itesm/pachontli/navigationDrawer/ui/Alimentos/AlimentosViewModel.kt
@@ -1,13 +1,13 @@
-package mx.itesm.pachontli.navigationDrawer.ui.home
+package mx.itesm.pachontli.navigationDrawer.ui.Alimentos
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class HomeViewModel : ViewModel() {
+class AlimentosViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
+        value = "This is gallery Fragment"
     }
     val text: LiveData<String> = _text
 }
