+++ app/src/main/java/mx/itesm/pachontli/navigationDrawer/ui/Interactuar/InteractuarFragment.kt
@@ -1,17 +1,17 @@
-package mx.itesm.pachontli.navigationDrawer.ui.home
+package mx.itesm.pachontli.navigationDrawer.ui.Interactuar
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import mx.itesm.pachontli.databinding.FragmentHomeBinding
+import mx.itesm.pachontli.databinding.FragmentInteractuarBinding
 
-class HomeFragment : Fragment() {
 
-    private var _binding: FragmentHomeBinding? = null
+class InteractuarFragment : Fragment() {
+
+    private var _binding: FragmentInteractuarBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +22,14 @@ class HomeFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val homeViewModel =
-            ViewModelProvider(this).get(HomeViewModel::class.java)
+        val slideshowViewModel =
+            ViewModelProvider(this).get(InteractuarViewModel::class.java)
 
-        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        _binding = FragmentInteractuarBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        /*val textView: TextView = binding.textHome
-        homeViewModel.text.observe(viewLifecycleOwner) {
+        /*val textView: TextView = binding.textSlideshow
+        slideshowViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }*/
         return root
