+++ app/src/main/java/mx/itesm/pachontli/navigationDrawer/ui/Veterinarios/VeterinariosFragment.kt
@@ -1,17 +1,16 @@
-package mx.itesm.pachontli.navigationDrawer.ui.slideshow
+package mx.itesm.pachontli.navigationDrawer.ui.Veterinarios
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import mx.itesm.pachontli.databinding.FragmentSlideshowBinding
+import mx.itesm.pachontli.databinding.FragmentVeterinariosBinding
 
-class SlideshowFragment : Fragment() {
+class VeterinariosFragment : Fragment() {
 
-    private var _binding: FragmentSlideshowBinding? = null
+    private var _binding: FragmentVeterinariosBinding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -22,14 +21,14 @@ class SlideshowFragment : Fragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        val slideshowViewModel =
-            ViewModelProvider(this).get(SlideshowViewModel::class.java)
+        val homeViewModel =
+            ViewModelProvider(this).get(VeterinariosViewModel::class.java)
 
-        _binding = FragmentSlideshowBinding.inflate(inflater, container, false)
+        _binding = FragmentVeterinariosBinding.inflate(inflater, container, false)
         val root: View = binding.root
 
-        /*val textView: TextView = binding.textSlideshow
-        slideshowViewModel.text.observe(viewLifecycleOwner) {
+        /*val textView: TextView = binding.textHome
+        homeViewModel.text.observe(viewLifecycleOwner) {
             textView.text = it
         }*/
         return root
