+++ app/src/main/java/mx/itesm/pachontli/navigationDrawer/ui/Veterinarios/VeterinariosViewModel.kt
@@ -1,13 +1,13 @@
-package mx.itesm.pachontli.navigationDrawer.ui.slideshow
+package mx.itesm.pachontli.navigationDrawer.ui.Veterinarios
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class SlideshowViewModel : ViewModel() {
+class VeterinariosViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is slideshow Fragment"
+        value = "This is home Fragment"
     }
     val text: LiveData<String> = _text
 }
