+++ app/src/main/res/drawable/side_nav_bar.xml
@@ -2,8 +2,8 @@
     android:shape="rectangle">
     <gradient
         android:angle="135"
-        android:centerColor="#009688"
-        android:endColor="#00695C"
-        android:startColor="#4DB6AC"
+        android:centerColor="#EE473D"
+        android:endColor="#FCBD29"
+        android:startColor="#FCBD29"
         android:type="linear" />
 </shape>
