+++ app/src/main/res/layout/activity_main.xml
@@ -21,9 +21,7 @@
             android:id="@+id/btnLogout"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="8dp"
-            android:layout_marginEnd="16dp"
+            android:layout_gravity="right"
             android:text="Cerrar Sesión" />
 
         <TextView
