+++ app/src/main/res/layout/app_bar_navigation_drawer.xml
@@ -15,7 +15,7 @@
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
             android:layout_height="?attr/actionBarSize"
-            android:background="?attr/colorPrimary"
+            android:background="#FCBD29"
             app:popupTheme="@style/Theme.Pachontli.PopupOverlay" />
 
     </com.google.android.material.appbar.AppBarLayout>
