+++ app/src/main/res/layout/fragment_alimentos.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".navigationDrawer.ui.home.HomeFragment">
+    tools:context=".navigationDrawer.ui.Veterinarios.VeterinariosFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
