+++ app/src/main/res/layout/fragment_interactuar.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FCBD29"
-    tools:context=".navigationDrawer.ui.slideshow.SlideshowFragment" >
+    tools:context=".navigationDrawer.ui.Interactuar.InteractuarFragment" >
 
     <TextView
         android:id="@+id/textView6"
