+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,19 +7,19 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="mx.itesm.pachontli.navigationDrawer.ui.home.HomeFragment"
+        android:name="mx.itesm.pachontli.navigationDrawer.ui.Veterinarios.VeterinariosFragment"
         android:label="@string/menu_home"
-        tools:layout="@layout/fragment_home" />
+        tools:layout="@layout/fragment_alimentos" />
 
     <fragment
-        android:id="@+id/nav_gallery"
-        android:name="mx.itesm.pachontli.navigationDrawer.ui.gallery.GalleryFragment"
+        android:id="@+id/nav_veterinarios"
+        android:name="mx.itesm.pachontli.navigationDrawer.ui.Alimentos.AlimentosFragment"
         android:label="@string/menu_gallery"
-        tools:layout="@layout/fragment_gallery" />
+        tools:layout="@layout/fragment_veterinarios" />
 
     <fragment
-        android:id="@+id/nav_slideshow"
-        android:name="mx.itesm.pachontli.navigationDrawer.ui.slideshow.SlideshowFragment"
+        android:id="@+id/nav_interactua"
+        android:name="mx.itesm.pachontli.navigationDrawer.ui.Interactuar.InteractuarFragment"
         android:label="@string/menu_slideshow"
-        tools:layout="@layout/fragment_slideshow" />
+        tools:layout="@layout/fragment_interactuar" />
 </navigation>
